MLOPS/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
gitpython==3.1.44
python-dotenv==1.1.1
numpy>=1.26.0
scipy>=1.11.4
